+++ quantify_scheduler/backends/zhinst/settings.py
@@ -12,7 +12,7 @@ from typing import Any, Callable, Dict, List, Tuple, Union, cast
 
 import numpy as np
 from zhinst.qcodes import base
-
+from quantify_core.utilities.general import make_hash
 from quantify_scheduler.backends.types import zhinst as zi_types
 from quantify_scheduler.backends.zhinst import helpers as zi_helpers
 
@@ -79,6 +79,14 @@ class ZISettings:
         self._daq_settings: List[ZISetting] = daq_settings
         self._awg_settings: List[Tuple[int, ZISetting]] = awg_settings
         self._awg_indexes = [awg_index for (awg_index, _) in self._awg_settings]
+
+    def __eq__(self, other):
+        self_dict = self.as_dict()
+        if not isinstance(other, ZISettings):
+            return False
+        other_dict = other.as_dict()
+        settings_equal = make_hash(self_dict) == make_hash(other_dict)
+        return settings_equal
 
     @property
     def awg_indexes(self) -> List[int]:
@@ -305,9 +313,17 @@ class ZISettingsBuilder:
     """
     The Zurich Instruments Settings builder class.
 
-    This class provides an API for settings that
-    are configured in the zhinst backend. The ZISettings
-    class is the resulting set that holds settings.
+    This class provides an API for settings that are configured in the zhinst backend.
+    The ZISettings class is the resulting set that holds settings.
+
+    This class exist because configuring these settings requires logic in how the
+    settings are configured using the zurich instruments API.
+
+    .. tip::
+
+        Build the settings using :meth:`~.build` and then view them as a dictionary
+        using :meth:`ZISettings.as_dict` to see what settings will be configured.
+
     """
 
     _daq_settings: List[ZISetting]
@@ -948,7 +964,7 @@ class ZISettingsBuilder:
         )
 
     def with_compiler_sourcestring(
-        self, awg_index: int, seqc: str, waveforms_dict: dict = None
+        self, awg_index: int, seqc: str
     ) -> ZISettingsBuilder:
         """
         Adds the sequencer compiler sourcestring
@@ -972,7 +988,6 @@ class ZISettingsBuilder:
                 partial(
                     zi_helpers.set_and_compile_awg_seqc,
                     awg_index=awg_index,
-                    waveforms_dict=waveforms_dict,
                 ),
             ),
         )