+++ quantify_scheduler/instrument_coordinator/components/zhinst.py
@@ -9,6 +9,7 @@ from __future__ import annotations
 
 import shutil
 from pathlib import Path
+import logging
 from typing import Dict, TYPE_CHECKING, Any, Optional, Tuple
 
 import numpy as np
@@ -22,6 +23,9 @@ if TYPE_CHECKING:
     from zhinst.qcodes.base import ZIBaseInstrument
     from quantify_scheduler.backends.zhinst_backend import ZIDeviceConfig
     from quantify_scheduler.backends.zhinst.settings import ZISettings
+
+
+logger = logging.getLogger(__name__)
 
 
 def convert_to_instrument_coordinator_format(acquisition_results):
@@ -45,7 +49,7 @@ class ZIInstrumentCoordinatorComponent(base.InstrumentCoordinatorComponentBase):
     def __init__(self, instrument: ZIBaseInstrument, **kwargs) -> None:
         """Create a new instance of ZIInstrumentCoordinatorComponent."""
         super().__init__(instrument, **kwargs)
-        self.device_config: Optional[ZIDeviceConfig] = None
+        self.zi_device_config: Optional[ZIDeviceConfig] = None
         self.zi_settings: Optional[ZISettings] = None
         self._data_path: Path = Path(".")
 
@@ -53,19 +57,42 @@ class ZIInstrumentCoordinatorComponent(base.InstrumentCoordinatorComponentBase):
     def is_running(self) -> bool:
         raise NotImplementedError()
 
-    def prepare(self, options: ZIDeviceConfig) -> None:
+    # pylint: disable=arguments-differ
+    def prepare(self, zi_device_config: ZIDeviceConfig) -> bool:
         """
         Prepare the InstrumentCoordinator component with configuration
         required to arm the instrument.
 
+        The preparation is skipped when the new zi_device_config is the same as that
+        from the previous time prepare was called. This saves significant time overhead.
+
         Parameters
         ----------
-        options :
-            The ZI instrument configuration.
-        """
-        self.device_config = options
+        zi_device_config :
+            The ZI instrument configuration. See the link for details of the
+            configuration format.
 
-        self.zi_settings = self.device_config.settings_builder.build()
+        Returns
+        -------
+        :
+            A boolean indicating if the ZI component was configured in this call.
+        """
+        self.zi_device_config = zi_device_config
+
+        new_zi_settings = zi_device_config.settings_builder.build()
+        old_zi_settings = self.zi_settings
+
+        if new_zi_settings == old_zi_settings:
+            logger.info(
+                f"{self.name}: device config and settings "
+                + "are identical! Compilation skipped."
+            )
+            return False
+
+        logger.info(f"Configuring {self.name}.")
+        # if the settings are not identical, update the attributes of the
+        # ic component and apply the settings to the hardware.
+        self.zi_settings = new_zi_settings
 
         # Writes settings to filestorage
         self._data_path = Path(handling.get_datadir())
@@ -78,6 +105,8 @@ class ZIInstrumentCoordinatorComponent(base.InstrumentCoordinatorComponentBase):
 
         # Upload settings, seqc and waveforms
         self.zi_settings.apply(self.instrument)
+
+        return True
 
     def retrieve_acquisition(self) -> Any:
         return None
@@ -128,9 +157,6 @@ class HDAWGInstrumentCoordinatorComponent(ZIInstrumentCoordinatorComponent):
         for awg_index in self.zi_settings.awg_indexes:
             self.get_awg(awg_index).stop()
 
-    def prepare(self, options: ZIDeviceConfig) -> None:
-        super().prepare(options)
-
     def retrieve_acquisition(self) -> Any:
         return None
 
@@ -161,29 +187,70 @@ class UHFQAInstrumentCoordinatorComponent(ZIInstrumentCoordinatorComponent):
     def stop(self) -> None:
         self.instrument.awg.stop()
 
-    def prepare(self, options: ZIDeviceConfig) -> None:
+    def prepare(self, zi_device_config: ZIDeviceConfig) -> bool:
         """
         Prepares the component with configurations
         required to arm the instrument.
+
         After this step is complete, the waveform file is uploaded
         to the LabOne WebServer.
+
+        Parameters
+        ----------
+        zi_device_config :
+            The ZI instrument configuration. See the link for details of the
+            configuration format.
+
+        Returns
+        -------
+        :
+            A boolean indicating if the ZI component was configured in this call.
         """
 
-        super().prepare(options)
-        self._data_path = Path(handling.get_datadir())
+        try:
+            # if settings where identical, no configuration is needed.
+            configure = super().prepare(zi_device_config)
+            if configure is False:
+                return False
 
+        # pylint: disable=broad-except
+        # the exception being raised is "Upload failed", but the ZI backend raises it
+        # as a general exception.
+        except Exception as e:
+            # whenever a new UHF device is used for the first time,
+            # certain waveform files will not exist. The lines below copy files so
+            # that it is possible to read from that location.
+            # this line of code should only be logging a warning the very first time
+            # a new setup is used, and then resolve auto.
+            logger.warning(e)
+            configure = True
+
+        self._data_path = Path(handling.get_datadir())
         # Copy the UHFQA waveforms to the waves directory
         # This is required before compilation.
+
+        # N.B. note this copies waves that were written during compilation, but are not
+        # contained in the zi_device_config that is passed as an argument here.
         waves_path: Path = zi_helpers.get_waves_directory(self.instrument.awg)
         wave_files = list(self._data_path.glob(f"{self.name}*.csv"))
         for file in wave_files:
             shutil.copy2(str(file), str(waves_path))
 
+        # prepare twice to resolve issue with waveform memory not being updated
+        # correctly. In practice, we see that integration weights update correctly, but
+        # the waveforms in pulses do not. This problem is not fully understood, but this
+        # resolves the issue at a minor overhead.
+
+        if configure:
+            # Upload settings, seqc and waveforms
+            self.zi_settings.apply(self.instrument)
+        return True
+
     def retrieve_acquisition(self) -> Dict[int, np.ndarray]:
-        if self.device_config is None:
+        if self.zi_device_config is None:
             raise RuntimeError("Undefined device config, first prepare UHFQA!")
 
-        acq_config = self.device_config.acq_config
+        acq_config = self.zi_device_config.acq_config
 
         acq_channel_results: Dict[int, np.ndarray] = dict()
         for acq_channel, resolve in acq_config.resolvers.items():